[FEATURE-13] Move entities to Entites folder
