+++ src/Host/ClientApp/src/entities/CounterState.ts
@@ -1,0 +1,3 @@
+export interface CounterState {
+    count: number;
+}
