+++ src/Host/ClientApp/src/entities/WeatherForecast.ts
@@ -1,0 +1,6 @@
+export interface WeatherForecast {
+    date: string;
+    temperatureC: number;
+    temperatureF: number;
+    summary: string;
+}
