+++ src/Host/ClientApp/src/features/Counter.ts
@@ -1,11 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-
-// -----------------
-// Состояние (state) - определяет тип данных, хранящихся в хранилище (store) Redux.
-
-interface CounterState {
-    count: number;
-}
+import { CounterState } from '../entities/CounterState';
 
 const initialState: CounterState = {
     count: 0
